+++ pocket/core/check.go
@@ -1,3 +1,16 @@
+// Copyright 2019 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package core
 
 import (
@@ -6,116 +19,61 @@ import (
 	"sort"
 	"strings"
 	"time"
-	"github.com/juju/errors"
 	"github.com/ngaut/log"
-	smith "github.com/pingcap/tipocket/go-sqlsmith"
-	"github.com/pingcap/tipocket/pocket/pkg/types"
+	"github.com/juju/errors"
 	"github.com/pingcap/tipocket/pocket/executor"
+	"github.com/pingcap/tipocket/pocket/pkg/types"
 	"github.com/pingcap/tipocket/pocket/util"
 )
 
-func (e *Executor) reloadSchema() error {
-	schema, err := e.coreConn.FetchSchema(e.dbname)
-	if err != nil {
-		return errors.Trace(err)
-	}
-	indexes := make(map[string][]string)
-	for _, col := range schema {
-		if _, ok := indexes[col[2]]; ok {
+func (c *Core) startCheckConsistency() {
+	t := 0
+	waitingForCheck := false
+	for range time.Tick(c.cfg.Options.CheckDuration.Duration) {
+		if waitingForCheck {
 			continue
 		}
-		index, err := e.coreConn.FetchIndexes(e.dbname, col[1])
-		// may not return error here
-		// just disable indexes
-		if err != nil {
-			return errors.Trace(err)
-		}
-		indexes[col[1]] = index
-	}
-
-	e.ss = smith.New()
-	e.ss.LoadSchema(schema, indexes)
-	e.ss.SetDB(e.dbname)
-	e.ss.SetStable(e.coreOpt.Stable)
-	return nil
-}
-
-func (e *Executor) startDataCompare() {
-	switch e.mode {
-	case "abtest":
-		e.startABTestDataCompare()
-	case "binlog":
-		e.startBinlogTestDataCompare()
-	default:
-		log.Fatalf("unhandled mode %s", e.mode)
-	}
-}
-
-func (e *Executor) startABTestDataCompare() {
-	c := time.Tick(time.Minute)
-	time := 0
-	for range c {
-		if e.waitingForCheck {
-			continue
-		}
-		e.waitingForCheck = true
-		time++
-		go func(time int) {
+		waitingForCheck = true
+		t++
+		go func(t int) {
 			log.Info("ready to compare data")
-			result, err := e.abTestCompareData(true)
-			log.Infof("test %d compare data result %t\n", time, result)
+			result, err := c.checkConsistency(false)
 			if err != nil {
+				// TODO: pass error by channel, stop process from outside
 				log.Fatalf("compare data error %+v", errors.ErrorStack(err))
 			}
-			e.waitingForCheck = false
-		}(time)
+			log.Infof("test %d compare data result %t\n", t, result)
+			waitingForCheck = false
+		}(t)
 	}
 }
 
-func (e *Executor) startBinlogTestDataCompare() {
-	c := time.Tick(time.Minute)
-	time := 0
-	for range c {
-		if e.waitingForCheck {
-			continue
-		}
-		e.waitingForCheck = true
-		time++
-		go func(time int) {
-			log.Info("ready to compare binlog data")
-			result, err := e.binlogTestCompareData(true)
-			log.Infof("test %d compare binlog data result %t\n", time, result)
-			if err != nil {
-				log.Fatalf("compare binlog data error %+v", errors.ErrorStack(err))
-			}
-			e.waitingForCheck = false
-		}(time)
-	}
-}
-
-func (e *Executor) checkConsistency(delay bool) (bool, error) {
-	switch e.mode {
+func (c *Core) checkConsistency(delay bool) (bool, error) {
+	var (
+		result bool
+		err    error
+	)
+	switch c.cfg.Mode {
 	case "abtest":
-		return e.abTestCompareData(delay)
+		result, err = c.abTestCompareData(delay)
 	case "binlog":
-		return e.binlogTestCompareData(delay)
+		result, err = c.binlogTestCompareData(delay)
+	default:
+		result, err = true, nil
 	}
-	// panic here?
-	return true, nil
-}
+	return result, errors.Trace(err)
+} 
 
 // abTestCompareDataWithoutCommit take snapshot without other transactions all committed
 // this function can run async and channel is for waiting taking snapshot complete
-func (e *Executor) abTestCompareDataWithoutCommit(ch chan struct{}) {
+func (c *Core) abTestCompareDataWithoutCommit(ch chan struct{}) {
 	// start a temp session for keep the snapshot of state
-	opt := e.execOpt.Clone()
-	opt.Mute = true
-	compareExecutor, err := executor.NewABTest(e.dsn1, e.dsn2, opt)
+	compareExecutor, err := c.initCompareConnection()
 	if err != nil {
 		log.Fatal(err)
 	}
 	// schema should be fetch first
-	schema, err := compareExecutor.GetConn().FetchSchema(e.dbname)
+	schema, err := compareExecutor.GetConn().FetchSchema(c.dbname)
 	if err != nil {
 		log.Fatal(err)
 	}
@@ -141,14 +99,14 @@ func (e *Executor) abTestCompareDataWithoutCommit(ch chan struct{}) {
 	log.Info("consistency check pass")
 }
 
-func (e *Executor) abTestCompareData(delay bool) (bool, error) {
+func (c *Core) abTestCompareData(delay bool) (bool, error) {
 	// only for abtest
-	if e.mode != "abtest" {
+	if c.cfg.Mode != "abtest" {
 		return true, nil
 	}
 
 	// start a temp session for keep the snapshot of state
-	compareExecutor, err := executor.NewABTest(e.dsn1, e.dsn2, e.execOpt.Clone())
+	compareExecutor, err := c.initCompareConnection()
 	if err != nil {
 		return false, errors.Trace(err)
 	}
@@ -160,18 +118,18 @@ func (e *Executor) abTestCompareData(delay bool) (bool, error) {
 
 	// commit or rollback all transactions
 	log.Info("before lock")
-	e.Lock()
+	c.Lock()
 	log.Info("after lock")
 	// no async here to ensure all transactions are committed or rollbacked in order
 	// use resolveDeadLock func to avoid deadlock
-	e.resolveDeadLock()
-	schema, err := compareExecutor.GetConn().FetchSchema(e.dbname)
+	c.resolveDeadLock()
+	schema, err := compareExecutor.GetConn().FetchSchema(c.dbname)
 	if err != nil {
-		e.Unlock()
+		c.Unlock()
 		return false, errors.Trace(err)
 	}
 	if err := compareExecutor.ABTestTxnBegin(); err != nil {
-		e.Unlock()
+		c.Unlock()
 		return false, errors.Trace(err)
 	}
 	if err := compareExecutor.ABTestSelect(makeCompareSQLs(schema)[0]); err != nil {
@@ -179,25 +137,26 @@ func (e *Executor) abTestCompareData(delay bool) (bool, error) {
 	}
 	// free the lock since the compare has already got the same snapshot in both side
 	// go on other transactions
-	// defer here for protect environment
-	defer e.Unlock()
+	// defer can be removed
+	// but here we use it for protect environment
+	defer c.Unlock()
 
 	// delay will hold on this snapshot and check it later
 	if delay {
 		time.Sleep(time.Duration(rand.Intn(5))*time.Second)
 	}
 	
-	return e.compareData(compareExecutor, schema)
+	return c.compareData(compareExecutor, schema)
 }
 
-func (e *Executor) binlogTestCompareData(delay bool) (bool, error) {
+func (c *Core) binlogTestCompareData(delay bool) (bool, error) {
 	// only for binlog test
-	if e.mode != "binlog" {
+	if c.cfg.Mode != "binlog" {
 		return true, nil
 	}
 
 	// start a temp session for keep the snapshot of state
-	compareExecutor, err := executor.NewABTest(e.dsn1, e.dsn2, e.execOpt.Clone())
+	compareExecutor, err := c.initCompareConnection()
 	if err != nil {
 		return false, errors.Trace(err)
 	}
@@ -207,20 +166,23 @@ func (e *Executor) binlogTestCompareData(delay bool) (bool, error) {
 		}
 	}(compareExecutor)
 
+	log.Info("before lock")
 	// commit or rollback all transactions
 	// lock here before get snapshot
-	e.Lock()
+	c.Lock()
+	log.Info("locked")
 	// no async here to ensure all transactions are committed or rollbacked in order
 	// use resolveDeadLock func to avoid deadlock
-	e.resolveDeadLock()
+	c.resolveDeadLock()
+	log.Info("resolve lock done")
 
 	// insert a table and wait for the sync job is done
 	table, tableStmt := generateWaitTable()
-	compareExecutor.SingleTestCreateTable(tableStmt)
+	compareExecutor.SingleTestExecDDL(tableStmt)
 	syncDone := false
 	for !syncDone {
 		time.Sleep(time.Second)
-		tables, err := compareExecutor.GetConn2().FetchTables(e.dbname)
+		tables, err := compareExecutor.GetConn2().FetchTables(c.dbname)
 		if err != nil {
 			return false, errors.Trace(err)
 		}
@@ -233,30 +195,31 @@ func (e *Executor) binlogTestCompareData(delay bool) (bool, error) {
 	}
 	time.Sleep(3 * time.Second)
 
-	schema, err := compareExecutor.GetConn().FetchSchema(e.dbname)
+	schema, err := compareExecutor.GetConn().FetchSchema(c.dbname)
 	if err != nil {
-		e.Unlock()
+		c.Unlock()
 		return false, errors.Trace(err)
 	}
 	if err := compareExecutor.ABTestTxnBegin(); err != nil {
-		e.Unlock()
+		c.Unlock()
 		return false, errors.Trace(err)
 	}
 	log.Info("compare wait for chan finish")
 	// free the lock since the compare has already got the same snapshot in both side
 	// go on other transactions
-	// defer here for protect environment
-	defer e.Unlock()
+	// defer can be removed
+	// but here we use it for protect environment
+	defer c.Unlock()
 
 	// delay will hold on this snapshot and check it later
 	if delay {
 		time.Sleep(time.Duration(rand.Intn(5))*time.Second)
 	}
 	
-	return e.compareData(compareExecutor, schema)
+	return c.compareData(compareExecutor, schema)
 }
 
-func (e *Executor) compareData(beginnedConnect *executor.Executor, schema [][5]string) (bool, error) {
+func (c *Core) compareData(beginnedConnect *executor.Executor, schema [][5]string) (bool, error) {
 	sqls := makeCompareSQLs(schema)
 	for _, sql := range sqls {
 		if err := beginnedConnect.ABTestSelect(sql); err != nil {
