+++ pocket/core/core.go
@@ -1,198 +1,72 @@
+// Copyright 2019 PingCAP, Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package core
 
 import (
 	"context"
-	"fmt"
-	"os"
-	"path"
-	"regexp"
 	"sync"
-
 	"github.com/juju/errors"
-	"github.com/ngaut/log"
-	"github.com/pingcap/tipocket/go-sqlsmith"
+	"github.com/pingcap/tipocket/pocket/config"
 	"github.com/pingcap/tipocket/pocket/connection"
 	"github.com/pingcap/tipocket/pocket/executor"
-	"github.com/pingcap/tipocket/pocket/pkg/logger"
 	"github.com/pingcap/tipocket/pocket/pkg/types"
 )
 
-var (
-	dbnameRegex      = regexp.MustCompile(`([a-z0-9A-Z_]+)$`)
-	schemaConnOption = connection.Option{
-		Mute: true,
-	}
-)
-
-type execSQL struct {
-	sql        *types.SQL
-	executorID int
-}
-
-// Executor struct
-type Executor struct {
-	dsn1            string
-	dsn2            string
-	coreOpt         *Option
-	execOpt         *executor.Option
-	executors       []*executor.Executor
-	ss              *sqlsmith.SQLSmith
-	ch              chan *execSQL
-	logger          *logger.Logger
-	dbname          string
-	mode            string
-	deadlockCh      chan int
-	order           *types.Order
-	ifLock          bool
-	mutex           sync.Mutex
-	waitingForCheck bool
+// Core is for random test scheduler
+type Core struct {
+	// global config
+	cfg *config.Config
 	// DSN here is for fetch schema only
 	coreConn *connection.Connection
 	coreExec *executor.Executor
+	// core properties
+	dbname      string
+	executors   []*executor.Executor
+	lockWatchCh chan int
+	order       *types.Order
+	// lock
+	mutex  sync.Mutex
+	ifLock bool
 }
 
-// New create Executor
-func New(dsn string, coreOpt *Option, execOpt *executor.Option) (*Executor, error) {
-	e := Executor{
-		dsn1:       dsn,
-		coreOpt:    coreOpt,
-		execOpt:    execOpt,
-		ch:         make(chan *execSQL, 1),
-		deadlockCh: make(chan int, 1),
-		order:      types.NewOrder(),
-		mode:       "single",
+// New creates a Core struct
+func New(cfg *config.Config) *Core {
+	return &Core{
+		cfg: cfg,
+		lockWatchCh: make(chan int),
+		order: types.NewOrder(),
 	}
-
-	if coreOpt.Reproduce != "" {
-		execOpt.Mute = true
-	}
-	log.Info("coreOpt.Concurrency is", coreOpt.Concurrency)
-	for i := 0; i < coreOpt.Concurrency; i++ {
-		opt := execOpt.Clone()
-		opt.ID = i + 1
-		opt.LogSuffix = fmt.Sprintf("-%d", i+1)
-		exec, err := executor.New(e.dsn1, opt)
-		if err != nil {
-			return nil, errors.Trace(err)
-		}
-		e.executors = append(e.executors, exec)
-	}
-	return e.init()
-}
-
-// NewABTest create abtest Executor
-func NewABTest(dsn1, dsn2 string, coreOpt *Option, execOpt *executor.Option) (*Executor, error) {
-	e := Executor{
-		dsn1:       dsn1,
-		dsn2:       dsn2,
-		coreOpt:    coreOpt,
-		execOpt:    execOpt,
-		ch:         make(chan *execSQL, 1),
-		deadlockCh: make(chan int, 1),
-		order:      types.NewOrder(),
-		mode:       "abtest",
-	}
-	if coreOpt.Mode != "" {
-		e.mode = coreOpt.Mode
-	}
-
-	if coreOpt.Reproduce != "" {
-		execOpt.Mute = true
-	}
-	log.Info("coreOpt.Concurrency is", coreOpt.Concurrency)
-	for i := 0; i < coreOpt.Concurrency; i++ {
-		opt := execOpt.Clone()
-		opt.ID = i + 1
-		opt.LogSuffix = fmt.Sprintf("-%d", i+1)
-		var (
-			exec *executor.Executor
-			err  error
-		)
-		switch e.mode {
-		case "abtest":
-			exec, err = executor.NewABTest(e.dsn1, e.dsn2, opt)
-		case "binlog":
-			exec, err = executor.New(e.dsn1, opt)
-		}
-		if err != nil {
-			return nil, errors.Trace(err)
-		}
-		e.executors = append(e.executors, exec)
-	}
-	return e.init()
-}
-
-func (e *Executor) init() (*Executor, error) {
-	// parse dbname
-	dbname := dbnameRegex.FindString(e.dsn1)
-	if dbname == "" {
-		return nil, errors.NotFoundf("empty dbname in dsn")
-	}
-	e.dbname = dbname
-	// init logger
-	l, err := logger.New("", false)
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	e.logger = l
-	// init schema exec
-	var exec *executor.Executor
-	switch e.mode {
-	case "single":
-		exec, err = executor.New(e.dsn1, e.execOpt.Clone())
-	case "abtest":
-		exec, err = executor.NewABTest(e.dsn1, e.dsn2, e.execOpt.Clone())
-	case "binlog":
-		exec, err = executor.New(e.dsn1, e.execOpt.Clone())
-	default:
-		panic("unhandled mode")
-	}
-	if err != nil {
-		return nil, errors.Trace(err)
-	}
-	e.coreExec = exec
-	e.coreConn = exec.GetConn()
-	return e, nil
 }
 
 // Start test
-func (e *Executor) Start(ctx context.Context) error {
-	if err := e.mustExec(); err != nil {
+func (c *Core) Start(ctx context.Context) error {
+	if err := c.initCoreConnectionWithoutSchema(); err != nil {
 		return errors.Trace(err)
 	}
-	if err := e.reloadSchema(); err != nil {
+	if err := c.prepare(); err != nil {
 		return errors.Trace(err)
 	}
-	for _, executor := range e.executors {
-		go executor.Start()
+	if err := c.initCoreConnection(); err != nil {
+		return errors.Trace(err)
 	}
-	if e.coreOpt.Reproduce != "" {
-		l, err := logger.New(path.Join(e.coreOpt.Reproduce, "reproduce.log"), false)
-		if err != nil {
-			log.Fatal(err)
-		}
-		e.logger = l
-		go e.reproduce()
-	} else {
-		go e.watchDeadLock()
-		go e.smithGenerate()
-		go e.startHandler()
-		switch e.mode {
-		case "abtest":
-			go e.startABTestDataCompare()
-		case "binlog":
-			go e.startBinlogTestDataCompare()
-		}
+	if err := c.initSubConnection(); err != nil {
+		return errors.Trace(err)
 	}
-
-	select {
-	case <-ctx.Done():
-		return nil
+	go c.watchLock()
+	if c.cfg.Options.Reproduce {
+		return c.reproduce(ctx)
 	}
-}
-
-// Stop test
-func (e *Executor) Stop(msg string) {
-	log.Infof("[STOP] message: %s\n", msg)
-	os.Exit(0)
+	go c.startCheckConsistency()
+	return errors.Trace(c.generate(ctx))
 }
