--- pocket/core/option.go
@@ -1,9 +1,0 @@
-package core
-
-// Option for core
-type Option struct {
-	Concurrency int
-	Stable      bool
-	Reproduce   string
-	Mode        string
-}
